added user dash board and some changes to staff
