+++ src/pages/Home/home.scss
@@ -101,3 +101,4 @@
 .home .dpts .dept{
   flex: 1  1 350px;
 }
+
