+++ src/pages/Staff/Staff.jsx
@@ -1,11 +1,27 @@
 import { ProfileCard } from "../../components"
 import './staff.scss';
-
+import { FaSearch} from 'react-icons/fa';
 const Staff = ()=>{
     return (
        <div className="staff">
             <div className="search">
-                <input type="text" placeholder="Search faculty"/>
+                <div className="input">
+                    <input type="text" placeholder="Search faculty"/><FaSearch className="search_btn" />
+                </div>
+                <div className="dropdown">
+                <select className="choices">
+                <option value="none" selected disabled hidden>Select an Option</option>
+                    <option >Teaching</option>
+                    <option>Non-Teaching</option>
+                </select>
+                <select className="choices">
+                <option value="none" selected disabled hidden>Select Subject</option>
+                    <option>Compiler Design</option>
+                    <option>Computer Networks</option>
+                    <option>DBMS</option>
+                </select>
+                </div>
+                
 
             </div>
             <div className="staff-row">
