+++ src/pages/Staff/staff.scss
@@ -3,17 +3,47 @@
         padding: 10px;
         display: flex;
         justify-content: center;
-        input {
-            padding: 10px;
+        .input {
+            padding: 10px 20px;
             margin: 20px;
+            
             margin-bottom: 30px;
             text-align: center;
             font-size: 15px;
             border: none;
             width: 300px;
             outline: none;
-            border-radius: 4px;
+            background-color: #fff;
+            border-radius: 18px;
+            display: flex;
+            align-items: center;
+            justify-content: space-between;
             box-shadow: 0 0 5px rgba(72, 72, 72, 0.12), 0px 2px 2px rgba(72, 72, 72, 0.5);
+            input{
+              background-color: none;
+              outline: none;
+              border: none;
+            }
+            .search_btn{
+              cursor: pointer;
+            }
+          }
+
+          .dropdown{
+            margin: 20px;
+            .choices{
+              margin: 4px;
+              margin-top: 2px;
+              border: none;
+              border-radius: 10px;
+              padding: 10px;
+              outline: none;
+              box-shadow: 0 0 5px rgba(72, 72, 72, 0.12), 0px 2px 2px rgba(72, 72, 72, 0.5);
+              
+              option{
+                color: black;
+              }
+            }
           }
       }
       .staff-row{
