+++ src/pages/user_Dash/EditProfile.scss
@@ -1,0 +1,125 @@
+.editprofile{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    .heading{
+        text-decoration-line: underline;
+        text-decoration-thickness: 3px;
+        font-size: 25px;
+        font-weight: 500;
+       
+    }
+    .image{
+        display: flex;
+        flex-direction: row;
+        padding: 10px;
+        margin: 20px;
+       
+        .edit_pic{
+            
+            display: flex;
+            flex-direction: column;
+             .edit_img{
+                
+                margin: 5px;
+            }
+            .profilepic{
+                height: 200px;
+                width: 200px;
+            }
+        }
+
+        .details{
+             .name{
+                border: 2px;
+                outline: none;
+                padding: 5px;
+                border-radius: 7px;
+                margin: 5px;
+             }
+
+            .subjects{
+                display: flex;
+                flex-direction: row;
+                .subject{
+                    border: 2px;
+                    outline: none;
+                    padding: 5px;
+                    border-radius: 7px;
+                    margin: 5px;
+                }
+                .add{
+                    border-radius: 12px;
+                    padding: 5px;
+                    color: #aaa;
+                    background-color:black;
+                    
+                    cursor: pointer;
+                }
+                
+            }
+
+            .icons{
+                margin-top: 15px;
+                input{
+                    margin: 5px;
+                    padding: 5px;
+                    border-radius: 5px;
+                }
+            }
+        }
+
+    }
+
+    .info{
+        min-width: 40%;
+        display: flex;
+        flex-direction: column;
+        
+        label{
+           
+            padding: 10px;
+            font-size: 20px;
+            font-style: oblique;
+            font-weight: 700;
+        }
+        .bio{
+            
+            padding: 5px;
+            
+            textarea{
+                width: 100%;
+                min-height: 100px;
+                padding: 5px;
+                outline: none;
+            }
+        }
+        .contact{
+            a{
+                color: blue;
+            }
+        }
+    }
+    .Buttons{
+        margin: 20px;
+        display: flex;
+        justify-content:end;
+        button{
+            padding: 10px;
+            margin: 5px;
+            padding-left: 25px;
+            padding-right: 25px;
+            background-color: green;
+            color: beige;
+            border-radius: 10px;
+        }
+
+        button:hover{
+            background-color: white;
+            border: 1px solid green;
+            color: green;
+        }
+    }
+
+}
